Move Opt, FileArg and Parent into a `cli` module
