+++ src/lib.rs
@@ -1,0 +1,3 @@
+extern crate structopt;
+
+pub mod cli;
